requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-nested-admin==3.2.4
django-nested-inlines==0.1
djangorestframework==3.12.4
future==0.18.2
html5lib==1.1
Pillow==9.0.0
psycopg2==2.9.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-monkey-business==1.0.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.5
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
xhtml2pdf==0.2.5
